refactor: move main file to cmd/bot and refactor by using handlers
